Refactored user determination on checks execution request
